Move game/parser/scanner to core/
