+++ core/scanner/scanner.go
